ml-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
mediapipe==0.10.14
opencv-python==4.11.0.86
numpy==1.24.4
scikit-learn==1.2.0
sse-starlette==1.6.5
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0